feat: Beginnings of testing refactor
