+++ tests/to_refactor/python/test_simulation_analysis.py
