+++ tests/to_refactor/python/test_varying_your_earnings.py
