+++ tests/to_refactor/python/test_yearly_var_removal.py
