+++ tests/unit/routes/test_error_routes.py
