+++ tests/unit/utils/payload_validators/test_validate_sim_analysis_payload.py
